refactor(bench): align Query workload layout
